chore: Refactor code to move Refit settings and client creation to ClickUpApi
In ServiceCollectionExtension.cs, we removed Refit settings instantiation and creation of IClickUpTeamApi client. Instead, HttpClient is now created and passed to ClickUpApi. In ClickUpApi.cs, the system is using HttpClient and instantiates the IClickUpTeamApi using RestService with relevant Refit settings. These changes make the ClickUpApi more responsible for its setup and configuration.