chore(WEBION-56): Refactor ClickUp integration to support v2 API and enhance authentication
This commit refactors ClickUp integration to support the v2 API. As part of the changes, the ClickUpAuthenticationOptions is removed, and the Webion.Stargaze.Auth.Handlers.ClickUp namespace is replaced with the Webion.AspNetCore.Authentication.ClickUp namespace. The ClickUpApiAuthHandler has been added to support token-based authentication, and the ClickUpHandler is added to the same namespace to enhance the authentication process. File renames and namespace changes align with the update to v2 API.