feat(WEBION-94): Add new controllers for Projects, Task, and TimeEntry management
A suite of new controllers has been added to facilitate the creation and management of projects, tasks, and time entries. These controllers include CRUD operations, as well as specific operations such as time entry synchronization and task billing. Also included in this commit is the establishment of package and rate-related endpoints for project time management.