feat(WEBION-109): add Tasks and TimePackages controllers (#3)
* feat(WEBION-109): add Task controller and setup auth

Co-authored-by: Stefano Calabretti <45517710+cala-br@users.noreply.github.com>

* feat(WEBION-109): add GetAllTimePackagesController

* feat(WEBION-109): add GetAllPackageRatesController

* fix(WEBION-109): improve validators

* fix(WEBION-109): improve AddPackageRateController validation

* refactor(WEBION-109): clean validators code

* fix(WEBION-109): improve CreateTimePackageController validation

* feat(WEBION-90): Add ClickUp schema to PostgreSQL database

This update includes the introduction of a new ClickUp schema to the PostgreSQL database. The migration file creates several new tables under the 'click_up' schema including 'space', 'folder', 'list' and 'task'. These additions allow for the incorporation of ClickUp data into the database structure.

* feat(WEBION-109): add RemovePackageRate and UpdatePackageRate controllers

* chore(WEBION-90): Add UserId and PackageId to TimePackageRateDto

UserId and PackageId fields have been added to the TimePackageRateDto and are now being populated in the GetAllPackageRatesController. This change allows API users to see which user and which package a specific rate is associated with.

---------

Co-authored-by: Stefano Calabretti <45517710+cala-br@users.noreply.github.com>
Co-authored-by: Stefano Calabretti <stefano.calabretti@webion.it>