feat: Remove ClickUp sync controllers and add refresh tokens to clients
Two controller classes for ClickUp synchronization (Folders and Lists), as part of the Stargaze API, were deleted. Furthermore, a migration was implemented to add refresh tokens to client functionality in the Stargaze PostgreSQL database. The latter change involves adding "clients_id" as a foreign key to the "refresh_token" table and setting up the corresponding constraints and relations.